requirements/dev.txt
-r ./base.txt

gunicorn==19.9.0
Werkzeug==0.14.1
ipdb==0.11
psycopg2==2.7.4 --no-binary psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.641
pytest==4.0.0
pytest-cov==2.7.0
pytest-django==3.4.4
pytest-sugar==0.9.2
responses==0.10.6
jsonschema==3.0.0


# Code quality
# ------------------------------------------------------------------------------
flake8==3.6.0
pylint==2.2.2
coverage==4.5.2

# Django
# ------------------------------------------------------------------------------
factory-boy==2.11.1
django-debug-toolbar==1.10.1
django-extensions==2.1.3

# Swagger
# ------------------------------------------------------------------------------
drf-yasg==1.11.0
flex==6.12.0
swagger-spec-validator==2.4.0
